requirements.txt
streamlit
google-generativeai
pypdf
langchain-community
duckduckgo-search
ddgs
pydantic<2.0
